wallet App

+++ src/main/java/com/masai/repository/BankAccountDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.masai.model.BankAccount;
 
 @Repository
-public interface BankAccountDao extends JpaRepository<BankAccount, Integer>{
+public interface BankAccountDao extends JpaRepository<BankAccount, Long>{
 
 }
